Added config and hooks to enable and disbale prometrics, moved help command code to commands and added metric hook to monitor
