+++ src/config/ProMetrics.ts
@@ -1,0 +1,3 @@
+export interface ProMetrics {
+    enabled: boolean;
+}
