+++ src/config/QMConfig.ts
@@ -5,6 +5,7 @@ import stripJsonComments = require("strip-json-comments");
 import {PrometheusClient} from "../gluon/metrics/prometheus/PrometheusClient";
 import {Cluster} from "./Cluster";
 import {HttpAuth} from "./HttpAuth";
+import {ProMetrics} from "./ProMetrics";
 import {SubatomicConfig} from "./SubatomicConfig";
 
 export class QMConfig {
@@ -19,6 +20,8 @@ export class QMConfig {
 
     public static cluster: Cluster;
 
+    public static proMetrics: ProMetrics;
+
     public static publicConfig() {
         return new PublicQMConfig();
     }
@@ -30,16 +33,20 @@ export class QMConfig {
         QMConfig.teamId = config.teamId;
         QMConfig.apiKey = config.apiKey;
         QMConfig.http = config.http;
+        QMConfig.proMetrics = config.proMetrics || {
+            enabled: true,
+        };
 
         QMConfig.cluster = config.cluster || {
             enabled: process.env.NODE_ENV === "production",
             workers: 10,
         };
-
-        if (QMConfig.cluster.enabled) {
-            QMConfig.http.customizers = [PrometheusClient.initializeClusteredMetricsServer];
-        } else {
-            QMConfig.http.customizers = [PrometheusClient.initializeNonClusteredMetricsServer];
+        if (QMConfig.proMetrics.enabled) {
+            if (QMConfig.cluster.enabled) {
+                QMConfig.http.customizers = [PrometheusClient.initializeClusteredMetricsServer];
+            } else {
+                QMConfig.http.customizers = [PrometheusClient.initializeNonClusteredMetricsServer];
+            }
         }
     }
