+++ src/gluon/util/shared/BaseQMCommand.ts
@@ -9,6 +9,7 @@ import {HandleCommand} from "@atomist/automation-client/lib/HandleCommand";
 import * as _ from "lodash";
 import {PrometheusClient} from "../../metrics/prometheus/PrometheusClient";
 import {BaseQMHandler} from "./BaseQMHandler";
+import {QMConfig} from "../../../config/QMConfig";
 
 export abstract class BaseQMComand extends BaseQMHandler implements HandleCommand<HandlerResult> {
 
@@ -21,23 +22,27 @@ export abstract class BaseQMComand extends BaseQMHandler implements HandleComman
     public abstract handle(ctx: HandlerContext);
 
     protected succeedCommand(message?: string) {
-        this.succeedHandler(message);
+        if (QMConfig.proMetrics.enabled) {
+            this.succeedHandler(message);
 
-        logger.debug(`teamChannel for prometheus ${this.teamChannel}`);
+            logger.debug(`teamChannel for prometheus ${this.teamChannel}`);
 
-        PrometheusClient.incrementCounter(`${_.snakeCase(this.constructor.name)}_command`, {
-            status: "success",
-            slackUsername: this.screenName,
-            team: this.teamChannel,
-        });
+            PrometheusClient.incrementCounter(`${_.snakeCase(this.constructor.name)}_command`, {
+                status: "success",
+                slackUsername: this.screenName,
+                team: this.teamChannel,
+            });
+        }
     }
 
     protected failCommand(message?: string) {
-        this.failHandler(message);
-        PrometheusClient.incrementCounter(`${_.snakeCase(this.constructor.name)}_command`, {
-            status: "fail",
-            slackUsername: this.screenName,
-            team: this.teamChannel,
-        });
+        if (QMConfig.proMetrics.enabled) {
+            this.failHandler(message);
+            PrometheusClient.incrementCounter(`${_.snakeCase(this.constructor.name)}_command`, {
+                status: "fail",
+                slackUsername: this.screenName,
+                team: this.teamChannel,
+            });
+        }
     }
 }
