+++ src/gluon/util/shared/BaseQMEvent.ts
@@ -1,15 +1,20 @@
 import * as _ from "lodash";
 import {PrometheusClient} from "../../metrics/prometheus/PrometheusClient";
 import {BaseQMHandler} from "./BaseQMHandler";
+import {QMConfig} from "../../../config/QMConfig";
 
 export abstract class BaseQMEvent extends BaseQMHandler {
 
     protected succeedEvent(message?: string) {
-        this.succeedHandler(message);
-        PrometheusClient.incrementCounter(`${_.snakeCase(this.constructor.name)}_event`, { status: "success"});
+        if (QMConfig.proMetrics.enabled) {
+            this.succeedHandler(message);
+            PrometheusClient.incrementCounter(`${_.snakeCase(this.constructor.name)}_event`, { status: "success"});
+        }
     }
     protected failEvent(message?: string) {
-        this.failHandler(message);
-        PrometheusClient.incrementCounter(`${_.snakeCase(this.constructor.name)}_event`, { status: "fail"});
+        if (QMConfig.proMetrics.enabled) {
+            this.failHandler(message);
+            PrometheusClient.incrementCounter(`${_.snakeCase(this.constructor.name)}_event`, { status: "fail"});
+        }
     }
 }
